did some stuff and added things. fixed program breaking bugs like the program just not writing stuff. it was great.

+++ ds_memory.h
@@ -39,4 +39,4 @@ long ds_write(long start, void *ptr, long bytes);
 
 int ds_finish();
 
-void seekPos(int pos);
+void seekPos(long pos);
